requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0.3
django-ckeditor==5.9.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-elasticsearch-dsl==7.1.1
django-js-asset==1.2.2
django-referrer-policy==1.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-mommy==2.0.0
mysqlclient==1.4.6
Pillow==7.0.0
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.9
wrapt==1.11.2
